ocr-service/requirements.txt
paddlepaddle==2.6.1
paddleocr==2.7.3
flask==3.0.0
flask-cors==4.0.0
opencv-python==4.9.0.80
Pillow==10.2.0
numpy==1.26.3
requests==2.31.0
gunicorn==21.2.0
